refactor: Move score type to a new file
